+++ test/runner/lib/delegation.py
@@ -12,15 +12,16 @@ import lib.thread
 
 from lib.executor import (
     SUPPORTED_PYTHON_VERSIONS,
+    create_shell_command,
+)
+
+from lib.config import (
+    TestConfig,
+    EnvironmentConfig,
     IntegrationConfig,
     ShellConfig,
     SanityConfig,
     UnitsConfig,
-    create_shell_command,
-)
-
-from lib.test import (
-    TestConfig,
 )
 
 from lib.core_ci import (
@@ -33,7 +34,6 @@ from lib.manage_ci import (
 
 from lib.util import (
     ApplicationError,
-    EnvironmentConfig,
     run_command,
     common_environment,
     pass_vars,
